外骨骼serial、socket库分开，添加读固件接口
