+++ pymycobot/__init__.py
@@ -42,7 +42,8 @@ from pymycobot.myarmm_control import MyArmMControl
 from pymycobot.mercurychassis_api import ChassisControl
 from pymycobot.conveyor_api import ConveyorAPI
 from pymycobot.ultraArmP340 import ultraArmP340
-from pymycobot.exoskeleton_api import Exoskeleton, ExoskeletonSocket
+from pymycobot.exoskeleton import Exoskeleton
+from pymycobot.exoskeletonsocket import ExoskeletonSocket
 
 __all__ = [
     "MyPalletizer260",
